requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.2
six==1.16.0
